Add files via upload
adding the remove column function. This is done by going through each row and then deleting the last cell in the row. Since add column adds a cell to the end of each row, remove column removes the last cell to keep it consistent. 

Also simply removed the console log of selectedColor since it is no longer needed after testing that the color does in deed change.