Add files via upload
adding the fill uncolored function. Similiar to the fill function it goes through an array of all the cells and changes the color of any with the default white to the color selected. Currently does not seem to work with initialized values, though it does work with cells after they have been "cleared". Upon testing and printing out the initalized value of background color does not seem to be white. will work on that with the addition of adding the click event