adding onclick & white color defaults
adding a helper function and its implementation to the creation of new cells. Upon initializing the cells add an onclick event to them that lets them change their background color to the color selected. Also initializing them with a white color helps with the previously bugged fill uncolored function testing